split global and task based content
